Moved the contract library and mocks to a contractLibrary folder.
