+++ source/services/dataContracts/resourceBuilder/resourceBuilder.service.ts
@@ -4,7 +4,7 @@ import * as angular from 'angular';
 
 import { IArrayUtility, serviceName as arrayServiceName, moduleName as arrayModuleName } from '../../array/array.service';
 
-import { IContractLibrary, ContractLibrary, ILibraryServices } from './contractLibrary';
+import { IContractLibrary, ContractLibrary, ILibraryServices } from '../contractLibrary/contractLibrary';
 import { IConverter } from '../baseDataServiceBehavior';
 import { IDataService, DataService, IBaseDomainObject } from '../dataService/data.service';
 import { IDataServiceView, IParentDataServiceView, DataServiceView, ParentDataServiceView } from '../dataService/dataServiceView';
